+++ lib/maze/hooks/appium_hooks.rb
@@ -45,8 +45,11 @@ module Maze
 
         start_driver(config, tunnel_id)
 
+        # Ensure the device is unlocked
+        Maze.driver.unlock
+
         # Write links to device farm sessions, where applicable
-        write_session_links
+        write_session_link
       end
 
       def before
@@ -88,16 +91,12 @@ module Maze
         end
       end
 
-      def write_session_links
+      def write_session_link
         config = Maze.config
         if config.farm == :bs && (config.device || config.browser)
           # Log a link to the BrowserStack session search dashboard
           build = Maze.driver.capabilities[:build]
-          url = if config.device
-                  "https://app-automate.browserstack.com/dashboard/v2/search?query=#{build}&type=builds"
-                else
-                  "https://automate.browserstack.com/dashboard/v2/search?query=#{build}&type=builds"
-                end
+          url = "https://app-automate.browserstack.com/dashboard/v2/search?query=#{build}&type=builds"
           if ENV['BUILDKITE']
             $logger.info Maze::LogUtil.linkify url, 'BrowserStack session(s)'
           else
@@ -160,6 +159,20 @@ module Maze
             config.capabilities = device_capabilities(config, tunnel_id)
             driver = create_driver(config)
             driver.start_driver unless config.appium_session_isolation
+
+            # Infer OS version if necessary when running locally
+            if Maze.config.farm == :local && Maze.config.os_version.nil?
+              version = case Maze.config.os
+              when 'android'
+                driver.session_capabilities['platformVersion'].to_f
+              when 'ios'
+                driver.session_capabilities['sdkVersion'].to_f
+              end
+              $logger.info "Inferred OS version to be #{version}"
+              Maze.config.os_version = version
+            end
+
+
             Maze.driver = driver
           rescue Selenium::WebDriver::Error::UnknownError => original_exception
             $logger.warn "Attempt to acquire #{config.device} device from farm #{config.farm} failed"